fix default build without -Poss-release
