+++ demos/client/src/main/java/README.txt
@@ -1,4 +1,4 @@
-SimpleClient
+example.client.SimpleClient
 ------------
 
 Notes:
@@ -27,12 +27,12 @@ To run the demo:
 
     4.  Run the demo. For example:
 
-	  java SimpleClient -L imap://username:password@hostname/
+	  java example.client.SimpleClient -L imap://username:password@hostname/
 
-	Note that SimpleClient expects to read the "simple.mailcap"
+	Note that example.client.SimpleClient expects to read the "simple.mailcap"
 	file from the current directory.  The simple.mailcap file
 	contains configuration information about viewers needed by
-	the SimpleClient demo program.
+	the example.client.SimpleClient demo program.
 
 
 
@@ -41,26 +41,26 @@ Overview of the Classes
 
 Main Classes:
 
-	SimpleClient   =    contains main().
+	example.client.SimpleClient   =    contains main().
 			     Uses the parameters to the application to
 			     locate the correct Store.  e.g.
 
-				SimpleClient -L imap://cotton:secret@snow-goon/
+				example.client.SimpleClient -L imap://cotton:secret@snow-goon/
 
 			     It will create the main frame and
 			     creates a tree.  The tree uses the
 			     StoreTreeNodes and FolderTreeNodes.
 
-	StoreTreeNode   =    subclass of Swing's DefaultMutableTreeNode.
+	example.client.StoreTreeNode   =    subclass of Swing's DefaultMutableTreeNode.
 			     This class shows how to get Folders from
 			     the Store.
 
-	FolderTreeNode  =    subclass of Swing's DefaultMutableTreeNode.
+	example.client.FolderTreeNode  =    subclass of Swing's DefaultMutableTreeNode.
 			     If the folder has messages, it will create
-			     a FolderViewer.  Otherwise it will add the
+			     a example.client.FolderViewer.  Otherwise it will add the
 			     subfolders to the tree.
 
-	SimpleAuthenticator = subclass of jakarta.mail.Authenticator. If
+	example.client.SimpleAuthenticator = subclass of jakarta.mail.Authenticator. If
 			     the Store is missing the username or the
 			     password, this authenticator will be used.
 			     It displays a dialog requesting the
@@ -69,22 +69,22 @@ Main Classes:
 
 Viewing Folders:
 
-	FolderViewer    =    Uses a Swing Table to display all of the
+	example.client.FolderViewer    =    Uses a Swing Table to display all of the
 			     Message in a Folder.  The "model" of the
-			     data for this Table is a FolderModel which
+			     data for this Table is a example.client.FolderModel which
 			     knows how to get displayable information
 			     from a Message.
 
 Jakarta Activation Viewers:
 
-	MessageViewer   =    Uses the content of the DataHandler.  The
+	example.client.MessageViewer   =    Uses the content of the DataHandler.  The
 			     content will be a jakarta.mail.Message
 			     object.  Displays the headers and then
 			     uses Jakarta Activation to find another viewer for
 			     the content type of the Message.  (either
 			     multipart/mixed, image/gif, or text/plain)
 
-	MultipartViewer =    Uses the content of the DataHandler.  The
+	example.client.MultipartViewer =    Uses the content of the DataHandler.  The
 			     content will be a jakarta.mail.Multipart
 			     object.  Uses Jakarta Activation to find another
 			     viewer for the first BodyPart's content.
@@ -93,9 +93,9 @@ Jakarta Activation Viewers:
 			     Button are pressed, it uses Jakarta Activation to
 			     find a viewer for the BodyPart's content,
 			     and displays it in a separate frame (using
-			     ComponentFrame).
+			     example.client.ComponentFrame).
 
-	TextViewer      =    Uses the content of the DataHandler.  The
+	example.client.TextViewer      =    Uses the content of the DataHandler.  The
 			     content will be either a java.lang.String
 			     object, or a java.io.InputStream object.
 			     Creates a TextArea and sets the text using
@@ -103,5 +103,5 @@ Jakarta Activation Viewers:
 
 Support Classes:
 
-	ComponentFrame  =    support class which takes a java.awt.Component
+	example.client.ComponentFrame  =    support class which takes a java.awt.Component
 			     and displays it in a Frame.
