+++ demos/servlet/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="iso-8859-1"?>
 <!--
 
-    Copyright (c) 1997, 2022 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 1997, 2023 Oracle and/or its affiliates. All rights reserved.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -30,39 +30,6 @@
     <modelVersion>4.0.0</modelVersion>
     <artifactId>servlet</artifactId>
     <name>Jakarta Mail API demo servlet</name>
-
-    <build>
-        <plugins>
-            <!--
-                Need to disable the maven-bundle-plugin because the
-                new version doesn't like classes in the default package.
-            -->
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>osgi-manifest</id>
-                        <phase>none</phase>
-                    </execution>
-                </executions>
-            </plugin>
-
-            <!--
-                Because the maven-jar-plugin depends on the manifest file
-                created by the maven-bundle-plugin, we need to disable it too.
-            -->
-            <plugin>
-                <artifactId>maven-jar-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>default-jar</id>
-                        <phase>none</phase>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
 
     <dependencies>
         <dependency>
