+++ doc/src/main/resources/docs/CHANGES.txt
@@ -7,6 +7,12 @@ for the Eclipse EE4J Angus Mail project:
 
     https://github.com/eclipse-ee4j/angus-mail/issues/<bug-number>
 
+          CHANGES IN THE 2.0.2 RELEASE
+          ----------------------------
+The following bugs have been fixed in the 2.0.2 release.
+
+89: fix default build without -Poss-release
+
           CHANGES IN THE 2.0.1 RELEASE
           ----------------------------
 The following bugs have been fixed in the 2.0.1 release.
